% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSEMs.R
\name{addSemMetadata}
\alias{addSemMetadata}
\title{Add metadata to SNPEffectMatrix}
\usage{
addSemMetadata(x, sem_metadata)
}
\arguments{
\item{x}{a SNPEffectMatrix object}

\item{sem_metadata}{a data.frame or named vector with metadata for the
given sem.}
}
\value{
SNPEffectMatrix object
}
\description{
Add metadata to SNPEffectMatrix
}
