% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{baseline}
\alias{baseline}
\alias{baseline,SNPEffectMatrix-method}
\title{Accessor function to pull the matrix slot from a SNPEffectMatrix object}
\usage{
baseline(x)

\S4method{baseline}{SNPEffectMatrix}(x)
}
\arguments{
\item{x}{a SNPEffectMatrix object}
}
\description{
Accessor function to pull the matrix slot from a SNPEffectMatrix object
}
