% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{calculatePvalues}
\alias{calculatePvalues}
\title{Permute sems for a given variant}
\usage{
calculatePvalues(permutation_counts, num_permutations)
}
\arguments{
\item{permutation_counts}{matrix storing permutation counts}

\item{num_permutations}{number of permutations to perform}
}
\description{
Permute sems for a given variant
}
