% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{cell_type,SemplScores-method}
\alias{cell_type,SemplScores-method}
\title{Accessor function to pull cell_type for a given sem id}
\usage{
\S4method{cell_type}{SemplScores}(x, s)
}
\arguments{
\item{x}{a SemplScores object}

\item{s}{a sem id represented in x}
}
\description{
Accessor function to pull cell_type for a given sem id
}
