% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{cellType}
\alias{cellType}
\alias{cellType,SNPEffectMatrix-method}
\title{Accessor function to pull the cell_type slot from a SNPEffectMatrix object}
\usage{
cellType(x)

\S4method{cellType}{SNPEffectMatrix}(x)
}
\arguments{
\item{x}{a SNPEffectMatrix object}
}
\description{
Accessor function to pull the cell_type slot from a SNPEffectMatrix object
}
