% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{changed_motif}
\alias{changed_motif}
\alias{changed_motif,data.table-method}
\title{Accessor function to subset the scores slot to changed motifs}
\usage{
changed_motif(scores_table, direction = "changed")

\S4method{changed_motif}{data.table}(scores_table, direction = "changed")
}
\arguments{
\item{scores_table}{the scores slot of a SemplScores object}

\item{direction}{direction of binding change. options are:
'changed', 'gained', 'lost', 'maintained'}
}
\description{
Accessor function to subset the scores slot to changed motifs
}
