% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_pos_to_vrange.R
\name{chr_pos_to_vrange}
\alias{chr_pos_to_vrange}
\title{Builds a VRanges object from genomic position and allele information}
\usage{
chr_pos_to_vrange(chr, pos, ref, alt)
}
\arguments{
\item{chr}{character, chromosome identifier}

\item{pos}{integer, SNP position on chromosome}

\item{ref}{character, reference nucleotide (can be upper or lower)}

\item{alt}{character, alternate nulceotide (can be upper or lower)}
}
\value{
a \code{VRanges} object the seqnames, ranges, ref, and alt columns
filled with parameter options supplied.
}
\description{
Builds a VRanges object from genomic position and allele information
}
