% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{countAlteredMotifs}
\alias{countAlteredMotifs}
\title{Permute sems for a given variant}
\usage{
countAlteredMotifs(permutation_counts, motif_count, sem)
}
\arguments{
\item{permutation_counts}{matrix storing permutation counts}

\item{motif_count}{matrix with a row for each motif and a column for each
motif change category (ie, gained, lost, changed, maintained)}

\item{sem}{unique sem id}
}
\description{
Permute sems for a given variant
}
