% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichment}
\alias{enrichment}
\title{Return a contingency table according to direction of test}
\usage{
enrichment(semScores, d = "changed")
}
\arguments{
\item{semScores}{a SemplScores object}

\item{d}{direction of test. Options are: "changed", "gained", "lost"}
}
\description{
Return a contingency table according to direction of test
}
