% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{ensembl}
\alias{ensembl}
\alias{ensembl,SNPEffectMatrix-method}
\title{Accessor function to pull the ensembl_id slot from a SNPEffectMatrix object}
\usage{
ensembl(x)

\S4method{ensembl}{SNPEffectMatrix}(x)
}
\arguments{
\item{x}{a SNPEffectMatrix object}
}
\description{
Accessor function to pull the ensembl_id slot from a SNPEffectMatrix object
}
