% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{ensembl_id}
\alias{ensembl_id}
\alias{ensembl_id,SemplScores-method}
\title{Accessor function to pull ensembl_id for a given sem id}
\usage{
ensembl_id(x, s)

\S4method{ensembl_id}{SemplScores}(x, s)
}
\arguments{
\item{x}{a SemplScores object}

\item{s}{a sem id represented in x}
}
\description{
Accessor function to pull ensembl_id for a given sem id
}
