% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{gainedContingencyTable}
\alias{gainedContingencyTable}
\title{Build a contingency table for gained motifs}
\usage{
gainedContingencyTable(s, si)
}
\arguments{
\item{s}{score table from a SemplScores object}

\item{si}{SEM id}
}
\description{
Build a contingency table for gained motifs
}
