% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{getContingencyTable}
\alias{getContingencyTable}
\title{Return a contingency table according to direction of test}
\usage{
getContingencyTable(s, si, d)
}
\arguments{
\item{s}{score table from a SemplScores object}

\item{si}{SEM id}

\item{d}{direction of test. Options are: "changed", "gained", "lost"}
}
\description{
Return a contingency table according to direction of test
}
