% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SNPEffectMatrix.R
\name{getSEM}
\alias{getSEM}
\alias{getSEM,SNPEffectMatrix-method}
\title{Access the SEM from a SNPEffectMatrix object}
\usage{
getSEM(x)

\S4method{getSEM}{SNPEffectMatrix}(x)
}
\arguments{
\item{x}{SNPEffectMatrix object}
}
\value{
A position (rows) by nucleic acid (columns)
\code{data.table} is returned
}
\description{
Access the SEM from a SNPEffectMatrix object
}
\examples{
## Create example SEM
df <- data.frame(A=c(1, 2, 3),
                 C=c(1, 2, 3),
                 G=c(1, 2, 3),
                 T=c(1, 2, 3))
s <- SNPEffectMatrix(df, 1.205, "motif_id")

## Access count matrix
getSEM(s)

}
