% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-SNPEffectMatrixCollection.R
\name{getSEMs}
\alias{getSEMs}
\alias{getSEMs,SNPEffectMatrixCollection-method}
\title{Access sems matrix from a SNPEffectMatrixCollection object}
\usage{
getSEMs(x, semId = NULL)

\S4method{getSEMs}{SNPEffectMatrixCollection}(x, semId = NULL)
}
\arguments{
\item{x}{SNPEffectMatrixCollection object}

\item{semId}{optional \code{character} corresponding to an SEM in the
SNPEffectMatrixCollection object. See semIds with \code{names(getSEMs(x))}.
Defaults to returning all SEMs.}
}
\value{
A position (rows) by nucleic acid (columns)
\code{data.table} is returned
}
\description{
Access sems matrix from a SNPEffectMatrixCollection object
}
\examples{
## Create example SEM
df <- data.frame(
    A = c(1, 2, 3),
    C = c(1, 2, 3),
    G = c(1, 2, 3),
    T = c(1, 2, 3)
)
s <- SNPEffectMatrix(df, 1.205, "motif_id")
sc <- SNPEffectMatrixCollection(s)

## Access count matrix
getSEMs(sc)

}
