% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSEMs.R
\name{loadSEM}
\alias{loadSEM}
\title{Load matrix and baseline data from .sem files}
\usage{
loadSEM(semFile, semId = NULL, bl = NULL)
}
\arguments{
\item{semFile}{A path to a .sem file. Expects header of file to be in format
\verb{#BASELINE:\{bl\}} where \code{bl} is the numeric baseline value. If matrix does
not include baseline header, must be specified in \code{bl}.}

\item{semId}{A unique id for the sem as a \code{character}. Defaults to
semFile file name without the extension.}

\item{bl}{\code{numeric} baseline value for the SEM. Overrides baseline specified
in semFile header.}
}
\value{
A SNPEffectMatrix object
}
\description{
Load matrix and baseline data from .sem files
}
