% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{metadata}
\alias{metadata}
\alias{metadata,SemplScores-method}
\title{Accessor metadata slot in a SemplScores object}
\usage{
metadata(x)

\S4method{metadata}{SemplScores}(x)
}
\arguments{
\item{x}{a SemplScores object}
}
\description{
Accessor metadata slot in a SemplScores object
}
