% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{motifScores,SemplScores-method}
\alias{motifScores,SemplScores-method}
\title{Accessor for sem binding scores for a motif}
\usage{
\S4method{motifScores}{SemplScores}(x, motif)
}
\arguments{
\item{x}{a SemplScores object}

\item{motif}{a sem id represented in x}
}
\description{
Accessor for sem binding scores for a motif
}
