% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{motifSub}
\alias{motifSub}
\alias{motifSub,SemplScores-method}
\title{Accessor function to subset a SemplScores object by a motif}
\usage{
motifSub(x, motif)

\S4method{motifSub}{SemplScores}(x, motif)
}
\arguments{
\item{x}{a SemplScores object}

\item{motif}{a sem id represented in x}
}
\description{
Accessor function to subset a SemplScores object by a motif
}
