% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permuteVariant}
\alias{permuteVariant}
\title{Permute sems for a given variant}
\usage{
permuteVariant(sem_scores, motif_count, variant_id)
}
\arguments{
\item{sem_scores}{SemplScores object}

\item{motif_count}{matrix with a row for each motif and a column for each
motif change category (ie, gained, lost, changed, maintained)}

\item{variant_id}{unique id for variant to be permuted}
}
\description{
Permute sems for a given variant
}
