% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSemEnrichment.R
\name{plotSemEnrichment}
\alias{plotSemEnrichment}
\title{Plot motif enrichment results}
\usage{
plotSemEnrichment(e, lab = "semId", sigThreshold = 0.05)
}
\arguments{
\item{e}{SEMplR enrichment results data.table}

\item{lab}{column in enrichment results to label significant points by.
Default is "semId".}

\item{sigThreshold}{adjusted pvalue threshold for coloring and labeling
points. Default is 0.05}
}
\value{
a \code{ggplot} with sem scores for each nucleic acid per position
}
\description{
Plot motif enrichment results
}
