% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSEMMotifs.R
\name{plotSEMMotifs}
\alias{plotSEMMotifs}
\title{Plot non-alt versus alt binding propensity for a single variant}
\usage{
plotSEMMotifs(
  s,
  variant,
  label = "transcription_factor",
  rc = TRUE,
  labsize = 4,
  cols = c("#F8766D", "dodgerblue2"),
  ptsize = 1
)
}
\arguments{
\item{s}{a SEMScores object with scores populated}

\item{variant}{variant id to plot}

\item{label}{column in sem_metadata slot of semplObj to use for point labels}

\item{rc}{label SEM orientations}

\item{labsize}{numeric size of the point labels}

\item{cols}{vector of length 2 with colors to use for plotting gained
and lost motifs respectively}

\item{ptsize}{numeric size of the ggplot points}
}
\value{
a \code{ggplot} scatter plot of non-alt binding propensity versus alt
binding propensity
}
\description{
Plot non-alt versus alt binding propensity for a single variant
}
\examples{
library(VariantAnnotation)
data(SEMC)

# create a VRanges object
vr <- VRanges(
    seqnames = "chr12",
    ranges = 94136009,
    ref = "G", alt = "C"
)

# calculate binding propensity
s <- scoreVariants(vr, SEMC, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)

plotSEMMotifs(s, "chr12:94136009:G>C", label = "transcription_factor")

}
