% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFlankingSeqs.R
\name{query_seqs}
\alias{query_seqs}
\title{Get up and downstream sequences of a variant}
\usage{
query_seqs(vr, up, down, bs_genome_obj = BSgenome.Hsapiens.UCSC.hg19::Hsapiens)
}
\arguments{
\item{vr}{A \code{VRanges} object with one or more variants. seqnames and ranges
fields are required.}

\item{up}{Numeric, number of bases to return upstream of variant}

\item{down}{Numeric, number of bases to return downstream of variant}

\item{bs_genome_obj}{A \code{BSgenome} object for the genome build to use.
Defaults to \code{BSgenome.Hsapiens.UCSC.hg19}.}
}
\value{
a named nested list with\code{upstream} and \code{downstream},
\code{ref_seq}, and \code{alt_seq} fields.
}
\description{
Finds nucleotides up and downstream of a variant within a genome
}
