% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFlankingSeqs.R
\name{querySeqs}
\alias{querySeqs}
\title{Get up and downstream sequences of a variant}
\usage{
querySeqs(x, up, down, bs_genome_obj, variant = T, allele = NULL)
}
\arguments{
\item{x}{A \code{VRanges} or \code{GRanges} object with one or more variants.
\code{seqnames} and \code{ranges} fields are required.}

\item{up}{Numeric, number of bases to return upstream of variant}

\item{down}{Numeric, number of bases to return downstream of variant}

\item{bs_genome_obj}{A \code{BSgenome} object for the genome build to use.
Defaults to \code{BSgenome.Hsapiens.UCSC.hg19}.}

\item{variant}{Boolean to indicate whether the provided query is a variant
or a single position.}

\item{allele}{If providing a \code{GRanges} object, can optionally provide an allele
to score for the indicated range. Otherwise, will query the BSgenome object
for the sequence at the indicated genomic position in x.}
}
\value{
a named nested list with\code{upstream}, \code{downstream},
\code{ref_seq}, and \code{alt_seq} fields when \code{x} is a \code{VRanges} object or just
\code{upstream}, \code{downstream}, and \code{seq} if \code{x} is a \code{GRanges} object.
}
\description{
Finds nucleotides up and downstream of a variant within a genome
}
