% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{scores,SemplScores-method}
\alias{scores,SemplScores-method}
\title{Accessor scores slot in a SemplScores object}
\usage{
\S4method{scores}{SemplScores}(x)
}
\arguments{
\item{x}{a SemplScores object}
}
\description{
Accessor scores slot in a SemplScores object
}
