% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SNPEffectMatrix.R
\name{sem}
\alias{sem}
\alias{sem,SNPEffectMatrix-method}
\title{Access sem matrix from a SNPEffectMatrix object}
\usage{
sem(x)

\S4method{sem}{SNPEffectMatrix}(x)
}
\arguments{
\item{x}{SNPEffectMatrix object}
}
\value{
A position (rows) by nucleic acid (columns)
\code{data.table} is returned
}
\description{
Access sem matrix from a SNPEffectMatrix object
}
\examples{
## Create example SEM
df <- data.frame(A=c(1, 2, 3),
                 C=c(1, 2, 3),
                 G=c(1, 2, 3),
                 T=c(1, 2, 3))
s <- SNPEffectMatrix(df, 1.205, "motif_id")

## Access count matrix
sem(s)

}
