% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{sem_metadata,SemplScores-method}
\alias{sem_metadata,SemplScores-method}
\title{Accessor sem_metadata slot in a SemplScores object}
\usage{
\S4method{sem_metadata}{SemplScores}(x)
}
\arguments{
\item{x}{a SemplScores object}
}
\description{
Accessor sem_metadata slot in a SemplScores object
}
