% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{semEnrichment}
\alias{semEnrichment}
\title{Perform a permuation test to find changed motifs among all variants analyzed}
\usage{
semEnrichment(sem_scores, n)
}
\arguments{
\item{sem_scores}{SemplScores object that has already been scored with
semMotifBinding}

\item{n}{number of permutations to run}
}
\value{
a data.frame with numbers of variants with altered motifs and
adjusted pvalues
}
\description{
Perform a permuation test to find changed motifs among all variants analyzed
}
