% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SNPEffectMatrix.R
\name{semId}
\alias{semId}
\alias{semId,SNPEffectMatrix-method}
\title{Access semId from a SNPEffectMatrix object}
\usage{
semId(x)

\S4method{semId}{SNPEffectMatrix}(x)
}
\arguments{
\item{x}{SNPEffectMatrix object}
}
\value{
The \code{character} semId is returned
}
\description{
Access semId from a SNPEffectMatrix object
}
\examples{
## Create example SEM
df <- data.frame(A=c(1, 2, 3),
                 C=c(1, 2, 3),
                 G=c(1, 2, 3),
                 T=c(1, 2, 3))
s <- SNPEffectMatrix(df, 1.205, "motif_id")

## Access count matrix
semId(s)

}
