% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semMotifBinding.R
\name{semMotifBinding}
\alias{semMotifBinding}
\title{Define a function to calculate risk/non-risk binding propensity
for all SEM motifs provided}
\usage{
semMotifBinding(vr, semList)
}
\arguments{
\item{vr}{\code{VRanges} object}

\item{semList}{a list of \code{SNPEffectMatrix} objects}
}
\value{
a SemplScores object
}
\description{
Define a function to calculate risk/non-risk binding propensity
for all SEM motifs provided
}
