% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SemplR-class}
\alias{SemplR-class}
\title{Class for storing SEM motif binding calculations for multiple variants}
\description{
Class for storing SEM motif binding calculations for multiple variants
}
\section{Slots}{

\describe{
\item{\code{variants}}{A \code{VRanges} object to hold one or more variants}

\item{\code{scores}}{A \code{data.table} object for motif information and binding scores}
}}

