% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMplR-package.R
\docType{package}
\name{SEMPLR-package}
\alias{SEMPLR}
\alias{SEMPLR-package}
\title{SEMPLR: SNP Effect Matrix Pipeline in R}
\description{
SEMPLR computes transcription factor binding affinity scores for genomic positions and genetic variants. Scores are computed from SNP Effect Matrices (SEMs) produced by SEMpl. 223 pre-computed SEMs are included with the package or custom sets can be provided. Enrichment can be tested among sets of genomic positions to determine if transcription factor binding events occur more often than expected. Comparing binding affinity scores between alleles can reveal differences in transcription factor binding with genetic variation. This package also includes several visualization functions to view scores both on the motif and variant/position level.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/grkenney/SEMPLR}
  \item \url{https://grkenney.github.io/SEMPLR}
  \item Report bugs at \url{https://www.github.com/grkenney/SEMPLR/issues}
}

}
\author{
\strong{Maintainer}: Grace Kenney \email{kenney.grace6@gmail.com} (\href{https://orcid.org/0009-0009-6308-3150}{ORCID})

Authors:
\itemize{
  \item Douglas Phanstiel
}

Other contributors:
\itemize{
  \item NIH NIGMS [funder]
  \item NSF GRFP [funder]
}

}
\keyword{internal}
