% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SemplR}
\alias{SemplR}
\alias{SemplR-class}
\title{SemplR object and constructor}
\usage{
SemplR(variants, scores)
}
\arguments{
\item{variants}{A \code{VRanges} object to hold one or more variants}

\item{scores}{A \code{data.table} object for motif information and binding scores}
}
\value{
a SemplR object
}
\description{
This constructor function would not typically be used by "end users".
This simple class indirectly extends the DataFrame class defined in the
S4Vectors package to allow other packages to write methods for results
objects from the DESeq2 package. It is used by \code{\link{results}}
to wrap up the results table.
}
