% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SEMplScores-class}
\alias{SEMplScores-class}
\title{Class for storing SEM motif binding calculations for multiple genomic ranges
or variants}
\description{
Class for storing SEM motif binding calculations for multiple genomic ranges
or variants
}
\section{Slots}{

\describe{
\item{\code{ranges}}{A \code{GRanges} or \code{VRanges} object to hold one or more genomic
ranges}

\item{\code{semData}}{A \code{data.table} object of metadata for the SEMs with one
row for each SEM}

\item{\code{scores}}{A \code{data.table} object for motif information and binding scores}
}}

