% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SequenceFrame-class}
\alias{SequenceFrame-class}
\title{Class to store information about sequencing frame and variant location}
\description{
Class to store information about sequencing frame and variant location
}
\section{Slots}{

\describe{
\item{\code{seqName}}{}

\item{\code{sequence}}{}
}}

