% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SequenceFrame}
\alias{SequenceFrame}
\title{SequenceFrame object and constructor}
\usage{
SequenceFrame(
  seqName,
  frameStart,
  frameStop,
  sequence,
  variantIndex,
  motif,
  tf,
  variantName
)
}
\arguments{
\item{seqName}{name to display alongside sequence}

\item{frameStart}{index of frame start}

\item{frameStop}{index of frame stop}

\item{sequence}{character sequence}

\item{variantIndex}{index of variant base pair(s)}

\item{motif}{sem id to display}

\item{tf}{tf name to display}

\item{variantName}{name of the variant to display}
}
\value{
a SequenceFrame object
}
\description{
Constructs a SequenceFrame class object.
}
