% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SEMplScores.R
\name{show,SEMplScores-method}
\alias{show,SEMplScores-method}
\title{Show method for SEMplScores objects}
\usage{
\S4method{show}{SEMplScores}(object)
}
\arguments{
\item{object}{a SEMplScores object}
}
\value{
An invisible NULL
}
\description{
Prints information about the number of variants, SEM meta data columns, and
the scoring table if scoreVariants has been run.
}
