% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SEMScores.R
\name{show,SEMScores-method}
\alias{show,SEMScores-method}
\title{Show method for SEMScores objects}
\usage{
\S4method{show}{SEMScores}(object)
}
\arguments{
\item{object}{a SEMScores object}
}
\value{
An invisible NULL
}
\description{
Prints information about the number of variants, SEM meta data columns, and
the scoring table if scoreVariants has been run.
}
