% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SNPEffectMatrix.R
\name{show,SNPEffectMatrix-method}
\alias{show,SNPEffectMatrix-method}
\title{Show for SNPEffectMatrix}
\usage{
\S4method{show}{SNPEffectMatrix}(object)
}
\arguments{
\item{object}{SNPEffectMatrix object.}
}
\value{
An invisible NULL
}
\description{
Show for SNPEffectMatrix
}
