% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SNPEffectMatrixCollection.R
\name{show,SNPEffectMatrixCollection-method}
\alias{show,SNPEffectMatrixCollection-method}
\title{Show method for SNPEffectMatrixCollection objects}
\usage{
\S4method{show}{SNPEffectMatrixCollection}(object)
}
\arguments{
\item{object}{a SNPEffectMatrixCollection object}
}
\value{
An invisible NULL
}
\description{
Prints information about the number of SEMs and meta data columns included
in the object.
}
