% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SNPEffectMatrixCollection.R,
%   R/methods-SemplScores.R, R/methods-SequenceFrame.R
\name{show,SNPEffectMatrixCollection-method}
\alias{show,SNPEffectMatrixCollection-method}
\alias{show,SemplScores-method}
\alias{show,SequenceFrame-method}
\title{Show method for SNPEffectMatrixCollection objects}
\usage{
\S4method{show}{SNPEffectMatrixCollection}(object)

\S4method{show}{SemplScores}(object)

\S4method{show}{SequenceFrame}(object)
}
\arguments{
\item{object}{a SequenceFrame object}
}
\description{
Prints information about the number of SEMs and meta data columns included
in the object.

Prints information about the number of variants, SEM meta data columns, and
the scoring table if scoreVariants has been run.
}
