% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SNPEffectMatrix}
\alias{SNPEffectMatrix}
\alias{SNPEffectMatrix-class}
\title{SNPEffectMatrix object and constructor}
\usage{
SNPEffectMatrix(matrix, tf_name, baseline, pwm_filename = "")
}
\arguments{
\item{matrix}{A \code{data.table} object to hold one or more variants}

\item{tf_name}{A \code{character} name for the transcription factor}

\item{baseline}{A \code{numeric} scrambled baseline, representing the binding
score of randomly scrambled kmers of the same length.}

\item{pwm_filename}{A \code{character} name of the original PWM file used to
generate the SEM}
}
\value{
a SNPEffectMatrix object
}
\description{
Constructs a SNPEffectMatrix class object.
}
