% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SNPEffectMatrixCollection-class}
\alias{SNPEffectMatrixCollection-class}
\title{A collection of SNPEffectMatrix objects and their corresponding meta data}
\description{
A collection of SNPEffectMatrix objects and their corresponding meta data
}
\section{Slots}{

\describe{
\item{\code{sems}}{A list of SNPEffectMatrix objects}

\item{\code{semData}}{A \code{data.table} with meta data on each SEM}

\item{\code{semKey}}{The name of a column in semData that matches the semIds in the
sems list as a \code{character}. If column entries have a .sem suffix, a new
column named SEM_KEY will be created without the .sem suffixes.}
}}

