% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{tallyAlteredMotifs}
\alias{tallyAlteredMotifs}
\title{Count the number of variants with altered motifs for each motif}
\usage{
tallyAlteredMotifs(sem_scores, pc)
}
\arguments{
\item{sem_scores}{SemplScores object}

\item{pc}{matrix storing permutation counts}
}
\value{
pc matrix with actual counts populated
}
\description{
Count the number of variants with altered motifs for each motif
}
