% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{tf}
\alias{tf}
\alias{tf,SNPEffectMatrix-method}
\title{Accessor function to pull the tf slot from a SNPEffectMatrix object}
\usage{
tf(x)

\S4method{tf}{SNPEffectMatrix}(x)
}
\arguments{
\item{x}{a SNPEffectMatrix object}
}
\description{
Accessor function to pull the tf slot from a SNPEffectMatrix object
}
