% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{tf_name}
\alias{tf_name}
\alias{tf_name,SemplScores-method}
\title{Accessor function to pull tf_name for a given sem id}
\usage{
tf_name(x, s)

\S4method{tf_name}{SemplScores}(x, s)
}
\arguments{
\item{x}{a SemplScores object}

\item{s}{a sem id represented in x}
}
\description{
Accessor function to pull tf_name for a given sem id
}
