% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{uniprot}
\alias{uniprot}
\alias{uniprot,SNPEffectMatrix-method}
\title{Accessor function to pull the uniprot slot from a SNPEffectMatrix object}
\usage{
uniprot(x)

\S4method{uniprot}{SNPEffectMatrix}(x)
}
\arguments{
\item{x}{a SNPEffectMatrix object}
}
\description{
Accessor function to pull the uniprot slot from a SNPEffectMatrix object
}
