% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{uniprot_id}
\alias{uniprot_id}
\alias{uniprot_id,SemplScores-method}
\title{Accessor function to pull uniprot_id for a given sem id}
\usage{
uniprot_id(x, s)

\S4method{uniprot_id}{SemplScores}(x, s)
}
\arguments{
\item{x}{a SemplScores object}

\item{s}{a sem id represented in x}
}
\description{
Accessor function to pull uniprot_id for a given sem id
}
