% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{variants,SemplScores-method}
\alias{variants,SemplScores-method}
\title{Accessor variants slot in a SemplScores object}
\usage{
\S4method{variants}{SemplScores}(x)
}
\arguments{
\item{x}{a SemplScores object}
}
\description{
Accessor variants slot in a SemplScores object
}
