% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{variantScores}
\alias{variantScores}
\alias{variantScores,SemplScores-method}
\title{Accessor for sem binding scores for a variant}
\usage{
variantScores(x, v)

\S4method{variantScores}{SemplScores}(x, v)
}
\arguments{
\item{x}{a SemplScores object}

\item{v}{a variant id represented in x}
}
\description{
Accessor for sem binding scores for a variant
}
