% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{variantSub}
\alias{variantSub}
\alias{variantSub,SemplScores-method}
\title{Accessor function to subset a SemplScores object by a variant}
\usage{
variantSub(x, v)

\S4method{variantSub}{SemplScores}(x, v)
}
\arguments{
\item{x}{a SemplScores object}

\item{v}{a variant id represented in x}
}
\description{
Accessor function to subset a SemplScores object by a variant
}
