% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewFrames.R
\name{viewFrames}
\alias{viewFrames}
\title{View the top scoring frames for a given variant and SEM}
\usage{
viewFrames(s, vid = NULL, sid = NULL, scoreIndex = NULL)
}
\arguments{
\item{s}{SemplScores object}

\item{vid}{variant id corresponding to the frame to view}

\item{sid}{SEM id corresponding to the frame to view}

\item{scoreIndex}{index of score to view. \code{vid} and \code{sid} takes prescedent
over this parameter.}
}
\value{
a SequenceFrame object
}
\description{
View the top scoring frames for a given variant and SEM
}
