% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSEMs.R
\name{addSemMetadata}
\alias{addSemMetadata}
\title{Add metadata to SNPEffectMatrix}
\usage{
addSemMetadata(sem, metadata)
}
\arguments{
\item{sem}{a SNPEffectMatrix object}

\item{metadata}{a data.frame or named vector with metadata for the given sem.}
}
\value{
SNPEffectMatrix object
}
\description{
Add metadata to SNPEffectMatrix
}
