% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SNPEffectMatrix.R
\name{baseline}
\alias{baseline}
\alias{baseline,SNPEffectMatrix-method}
\title{Access baseline from a SNPEffectMatrix object}
\usage{
baseline(x)

\S4method{baseline}{SNPEffectMatrix}(x)
}
\arguments{
\item{x}{SNPEffectMatrix object}
}
\value{
The \code{numeric} baseline value is returned
}
\description{
Access baseline from a SNPEffectMatrix object
}
\examples{
## Create example SEM
df <- data.frame(A=c(1, 2, 3),
                 C=c(1, 2, 3),
                 G=c(1, 2, 3),
                 T=c(1, 2, 3))
s <- SNPEffectMatrix(df, 1.205, "motif_id")

## Access count matrix
baseline(s)

}
