% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSEMsToPPMs.R
\name{convertSEMsToPPMs}
\alias{convertSEMsToPPMs}
\title{Convert a SNP Effect Matrix to a position probability matrices (PPMs)}
\usage{
convertSEMsToPPMs(x)
}
\arguments{
\item{x}{A \code{SNPEffectMatrix} or \code{SNPEffectMatrixCollection} object}
}
\value{
a \code{list} of matrices
}
\description{
Converts the SNP Effect Matrix in a SNPEffectMatrix or
SNPEffectMatrixCollection object to a position probability matrix (PPM)
where each entry is the probability of a base at each position in the motif.
}
\examples{
# Load default SEMs
data(SEMC)

# From a SNPEffectMatrixCollection
convertSEMsToPPMs(SEMC)

# From a single SNPEffectMatrix
convertSEMsToPPMs(getSEMs(SEMC, "AP2B_HUMAN.SK-N-SH"))

}
