% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichSEMs.R
\name{enrichSEMs}
\alias{enrichSEMs}
\title{Calculates binding enrichment for motif(s)}
\usage{
enrichSEMs(x, semList, bs_genome_obj, background = NULL)
}
\arguments{
\item{x}{A \code{SNPEffectMatrix} or \code{SNPEffectMatrixCollection}}

\item{semList}{A \code{SNPEffectMatrix} or \code{SNPEffectMatrixCollection} object}

\item{bs_genome_obj}{A \code{BSgenome} object for the genome build to use.}

\item{background}{A list of DNA sequences to use for background. The length of each
sequence must match the length of sequences in \code{x}}
}
\value{
a \code{list} of matrices
}
\description{
Perform a binomial test to determine if SNP Effect Matrices are bound more
often than expected.
}
