% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SNPEffectMatrix.R
\name{getBaseline}
\alias{getBaseline}
\alias{getBaseline,SNPEffectMatrix-method}
\title{Access baseline from a SNPEffectMatrix object}
\usage{
getBaseline(x)

\S4method{getBaseline}{SNPEffectMatrix}(x)
}
\arguments{
\item{x}{SNPEffectMatrix object}
}
\value{
The \code{numeric} baseline value is returned
}
\description{
Access baseline from a SNPEffectMatrix object
}
\examples{
# Isolate a single SNPEffectMatrix object from the default
# SNPEffectMatrixCollection
sm <- getSEMs(SEMC)[[1]]

# Access the baseline
getBaseline(sm)

}
