% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoordinates.R
\name{getCoordinates}
\alias{getCoordinates}
\title{Retrieve Promoter Regions and Sample Background Elements}
\usage{
getCoordinates(
  mapped,
  txdb,
  transcript = FALSE,
  n_ratio = 1,
  promoterWindow = c(upstream = 300, downstream = 50),
  standardChroms = TRUE,
  reduceOverlaps = TRUE,
  overlapMinGap = 0,
  onePromoterPerGene = FALSE
)
}
\arguments{
\item{mapped}{A list as returned by \code{\link{mapIDs}()}, containing
at least \code{fg_ids} and \code{bg_ids}.}

\item{txdb}{A TxDb object.
(e.g. \code{TxDb.Hsapiens.UCSC.hg38.knownGene}).}

\item{transcript}{Logical; \code{TRUE} for transcript-level coordinates,
\code{FALSE} for gene-level.}

\item{n_ratio}{Numeric; ratio of ranges to retain in the background set.
The number of background ranges will be equal to \code{n_ratio} multiplied
by the number of foreground ranges.}

\item{promoterWindow}{Numeric named vector of lengths:
\code{c(upstream, downstream)} (default \code{c(300,50)}).}

\item{standardChroms}{Logical; restrict to standard chromosomes.}

\item{reduceOverlaps}{Logical; merge any overlapping promoter windows.}

\item{overlapMinGap}{Numeric; minimum gap when reducing overlaps.}

\item{onePromoterPerGene}{Logical; if \code{TRUE}, choose one
promoter per gene.}
}
\value{
A named \code{list} from the final
\code{.defineBackgroundElements()}:
\item{backgroundElements}{\code{GRanges} of sampled background promoters}
\item{foregroundElements}{\code{GRanges} of foreground promoters}
\item{backgroundUniverse}{\code{GRanges} of the pruned universe}
\item{matchObject}{\code{MatchedGRanges} if \code{bgMethod="matched"},
else \code{NULL}}
}
\description{
Given a mapped set of foreground IDs (with Entrez and mappedID),
this function:
\enumerate{
\item Fetches genomic coordinates for each feature (gene or transcript),
according to the chosen \code{TSS.method}.
\item Extracts promoter windows around those coordinates.
\item Optionally reduces overlaps and selects one promoter per gene.
\item Samples background promoter elements matching the foreground
}
}
\examples{
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
library(org.Hs.eg.db)

txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene
orgdb <- org.Hs.eg.db

my_genes <- c("ENSG00000139618", "ENSG00000157764")
ids <- mapIDs(orgdb = orgdb, 
              foreground_ids = my_genes, 
              id_type = "ENSEMBL")
filtered <- poolFilter(ids, geneType="protein-coding")
coords <- getCoordinates(mapped = filtered, txdb = txdb)

}
