% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFlankingSeqs.R
\name{getFlankingSeqs}
\alias{getFlankingSeqs}
\title{Get up and downstream sequences of variant(s) and store in VRanges metadata}
\usage{
getFlankingSeqs(
  vr,
  up,
  down,
  bs_genome_obj = BSgenome.Hsapiens.UCSC.hg19::Hsapiens
)
}
\arguments{
\item{vr}{A \code{VRanges} object with one or more variants. seqnames and ranges
fields are required.}

\item{up}{Numeric, number of bases to return upstream of variant}

\item{down}{Numeric, number of bases to return downstream of variant}

\item{bs_genome_obj}{A \code{BSgenome} object for the genome build to use.
Defaults to \code{BSgenome.Hsapiens.UCSC.hg19}.}
}
\value{
a \code{VRanges} object with metadata columns \code{upstream} and \code{downstream}.
}
\description{
Finds nucleotides up and downstream of a variant within a genome and store
within \code{upstream} and \code{downstream} fields in the metadata of the supplied
\code{VRanges} object.
}
