% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SEMScores.R
\name{getRanges}
\alias{getRanges}
\alias{getRanges,SEMScores-method}
\title{Access ranges slot in a SEMScores object}
\usage{
getRanges(x)

\S4method{getRanges}{SEMScores}(x)
}
\arguments{
\item{x}{a SEMScores object}
}
\value{
A GRanges or VRanges object
}
\description{
Access ranges slot in a SEMScores object
}
\examples{
library(VariantAnnotation)

# load default SEMs

# create a VRanges object
vr <- VRanges(
    seqnames = "chr12",
    ranges = 94136009,
    ref = "G", alt = "C"
)

# calculate binding propensity
s <- scoreVariants(vr, SEMC, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)

getRanges(s)

}
