% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SNPEffectMatrix.R
\name{getSEM}
\alias{getSEM}
\alias{getSEM,SNPEffectMatrix-method}
\title{Access the SEM from a SNPEffectMatrix object}
\usage{
getSEM(x)

\S4method{getSEM}{SNPEffectMatrix}(x)
}
\arguments{
\item{x}{SNPEffectMatrix object}
}
\value{
A position (rows) by nucleic acid (columns)
\code{data.table} is returned
}
\description{
Access the SEM from a SNPEffectMatrix object
}
\examples{
# Isolate a single SNPEffectMatrix object from the default 
# SNPEffectMatrixCollection
sm <- sems(sc)[[1]]

# Access the matrix
getSEM(sm)

}
