% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SNPEffectMatrix.R
\name{getSEMId}
\alias{getSEMId}
\alias{getSEMId,SNPEffectMatrix-method}
\title{Access semId from a SNPEffectMatrix object}
\usage{
getSEMId(x)

\S4method{getSEMId}{SNPEffectMatrix}(x)
}
\arguments{
\item{x}{SNPEffectMatrix object}
}
\value{
The \code{character} id is returned
}
\description{
Access semId from a SNPEffectMatrix object
}
\examples{
# Isolate a single SNPEffectMatrix object from the default 
# SNPEffectMatrixCollection
sm <- sems(sc)[[1]]

# Access the SEM id
getSEMId(sm)

}
