% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSEMs.R
\name{loadSEMs}
\alias{loadSEMs}
\title{Define a function to calculate risk/non-risk binding propensity}
\usage{
loadSEMs(semFiles)
}
\arguments{
\item{semFiles}{character vector of paths to each \code{.sem} file}
}
\description{
Define a function to calculate risk/non-risk binding propensity
}
