% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSEMs.R
\name{loadSEMs}
\alias{loadSEMs}
\title{Load .sem and baseline data from files or from github if file path is not
provided. Github default data: github.com/Boyle-Lab/SEMpl/raw/master/SEMs/}
\usage{
loadSEMs(sem_dir = NULL, baseline_file = NULL)
}
\arguments{
\item{sem_dir}{path to directory containing .sem files.
Defaults to pulling data from github.}

\item{baseline_file}{path to baselines file.
Defaults to pulling data from github.}
}
\description{
Load .sem and baseline data from files or from github if file path is not
provided. Github default data: github.com/Boyle-Lab/SEMpl/raw/master/SEMs/
}
