% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapIDs.R
\name{mapIDs}
\alias{mapIDs}
\title{Map User IDs to Entrez and Determine Best Keytype}
\usage{
mapIDs(
  orgdb,
  id_type,
  foreground_ids,
  background_ids = NULL,
  threshold = 0.9,
  transcript = FALSE,
  stripVersions = TRUE,
  inflateThresh = 1
)
}
\arguments{
\item{orgdb}{An OrgDb object. (e.g. \code{org.Hs.eg.db}).}

\item{id_type}{Type of identifier supplied in foreground and background IDs.}

\item{foreground_ids}{Character vector of gene or
transcript IDs (e.g. Ensembl, RefSeq, gene symbols) to analyze.}

\item{background_ids}{Character vector of gene or transcript
IDs to use as background set.}

\item{threshold}{Fraction in range 0 to 1; minimum mapping rate to
accept a keytype without falling back (default 0.9).}

\item{transcript}{Logical; if \code{TRUE}, analyze as transcript-level
IDs (default \code{FALSE}).}

\item{stripVersions}{Logical; strip version suffixes (e.g. ".1") from
Ensembl/RefSeq IDs.}

\item{inflateThresh}{Fraction in range 0 to 1; if reverse-mapping shows
excessive, inflation automatically collapse transcripts to genes
(default 1 ie. 100\%).}
}
\value{
A named list combining the original \code{mapping} components with:
\describe{
\item{\code{fg_ids}}{data.frame(entrez, mappedID) for your foreground set}
\item{\code{bg_ids}}{data.frame(entrez, mappedID) for the full background}
\item{\code{userIDtype}}{the chosen keytype (e.g. "ensembl")}
\item{\code{transcript}}{logical, whether transcript-level mapping was used}
}
}
\description{
Given a vector of user-supplied gene/transcript IDs, finds the AnnotationDbi
keytype (e.g. ens, refseq, symbol, etc.) that maps the highest fraction of
inputs, and returns both foreground and full background sets as Entrez IDs.
Optionally collapses transcript-style inputs to genes when requested or when
reverse-mapping inflation exceeds a threshold.
}
\examples{
library(org.Hs.eg.db)
orgdb <- org.Hs.eg.db

my_genes <- c("ENSG00000139618", "ENSG00000157764")
ids <- mapIDs(orgdb = orgdb, 
              foreground_ids = my_genes, 
              id_type = "ENSEMBL")
  
# Transcript Ids
my_transcripts <- c("ENST00000245479", "ENST00000633194")
ids <- mapIDs(orgdb = orgdb,
              foreground_ids = my_transcripts,
              id_type = "ENSEMBLTRANS",
              transcript = TRUE)

}
