% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrich.R
\name{plotEnrich}
\alias{plotEnrich}
\title{Plot the results of \code{enrichSEMs}}
\usage{
plotEnrich(e, semList, threshold = 0.05)
}
\arguments{
\item{e}{The resulting data.table from \code{enrichSEMs}}

\item{semList}{A \code{SNPEffectMatrixCollection} object}

\item{threshold}{The adjusted p-value threshold for coloring SEMs}
}
\value{
a \code{ggtree} object
}
\description{
Generates a cladogram clustered by SNP Effect Matrix similarity and
a heatmap representing the adjusted p-value of the enrichment.
}
