% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrich.R
\name{plotEnrich}
\alias{plotEnrich}
\title{Plot the results of \code{enrichSEMs}}
\usage{
plotEnrich(
  e,
  sem,
  label = "transcription_factor",
  method = "WPCC",
  threshold = 0.05,
  sigCol = "purple4"
)
}
\arguments{
\item{e}{The resulting data.table from \code{enrichSEMs}}

\item{sem}{A \code{SNPEffectMatrixCollection} object}

\item{label}{Column in semData(sem) to use for tree labels}

\item{method}{Method to use for SEM comparison.
See ?universalmotif::compare_motifs for options.}

\item{threshold}{The adjusted p-value threshold for coloring SEMs}

\item{sigCol}{Color to label significant SEM labels}
}
\value{
a \code{ggtree} object
}
\description{
Generates a cladogram clustered by SNP Effect Matrix similarity and
a heatmap representing the adjusted p-value of the enrichment.
}
