% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMotif.R
\name{plotMotif}
\alias{plotMotif}
\title{Plot motif SEM scores}
\usage{
plotMotif(sems, motif)
}
\arguments{
\item{sems}{list of SNPEffectMatrix objects}

\item{motif}{sem id to plot}
}
\value{
a \code{ggplot} with sem scores for each amino acid per position
}
\description{
Plot motif SEM scores
}
