% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSEM.R
\name{plotSEM}
\alias{plotSEM}
\title{Plot SEM scores for each nucleic acid in each position of the motif}
\usage{
plotSEM(
  sem,
  motif = NULL,
  motifSeq = NULL,
  highlight = NULL,
  hcol = "dodgerblue",
  hwidth = 15,
  size = 7
)
}
\arguments{
\item{sem}{A SNPEffectMatrix or SNPEffectMatrixCollection object.}

\item{motif}{sem id to plot. If providing a SNPEffectMatrixCollection object,
must match a name in the list. If providing a single SNPEffectMatrix object,
this parameter is ignored and the SNPEffectMatrix's semId is used for
plotting.}

\item{motifSeq}{Character sequence to color on plot}

\item{highlight}{Index of variant location on reference}

\item{hcol}{Color of motifSeq bases}

\item{hwidth}{Width of highlight bar}

\item{size}{Font size of nucleotides in plot}
}
\value{
a \code{ggplot} with sem scores for each nucleic acid per position
}
\description{
Plot SEM scores for each nucleic acid in each position of the motif
}
\examples{
library(VariantAnnotation)

# create an SNP Effect Matrix (SEM)
sem <- matrix(rnorm(12), ncol = 4)
colnames(sem) <- c("A", "C", "G", "T")

# create a list of SNPEffectMatrix objects
s <- SNPEffectMatrix(sem, baseline = -1, semId = "sem_id")

# plot the motif
plotSEM(s, getSEMId(s))

# color by sequence
plotSEM(s, getSEMId(s), motifSeq = "ACT", highlight = 2)
}
