% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSEM.R
\name{plotSEM}
\alias{plotSEM}
\title{Plot SEM scores for each nucleic acid in each position of the motif}
\usage{
plotSEM(
  sem,
  motif = NULL,
  motifSeq = NULL,
  cols = c("lightgrey", "dodgerblue"),
  size = 7,
  hindex = NULL,
  hcol = "dodgerblue",
  halpha = 0.1,
  hwidth = 15,
  lcol = "#d7dbdd",
  lwidth = 1
)
}
\arguments{
\item{sem}{A SNPEffectMatrix or SNPEffectMatrixCollection object.}

\item{motif}{sem id to plot. If providing a SNPEffectMatrixCollection object,
must match a name in the list. If providing a single SNPEffectMatrix object,
this parameter is ignored and the SNPEffectMatrix's semId is used for
plotting.}

\item{motifSeq}{Character sequence to color on plot}

\item{cols}{A vector of two colors to color the nucleotides not included
and included in the provided motifSeq respectively}

\item{size}{Font size of nucleotides in plot}

\item{hindex}{Index of nucleotide to highlight in the motifSeq}

\item{hcol}{The color of the vertical highlight bar}

\item{halpha}{The alpha transparency of the vertical highlight bar}

\item{hwidth}{Width of highlight bar}

\item{lcol}{Color of the horizongtal endogenous and background binding lines}

\item{lwidth}{Width of the horizontal endogenous and background binding lines}
}
\value{
a \code{ggplot} with sem scores for each nucleic acid per position
}
\description{
Plot SEM scores for each nucleic acid in each position of the motif
}
\examples{
library(VariantAnnotation)
data(SEMC)

# Given a SNPEffectMatrix Collection
plotSEM(SEMC, motif = "JUN")

# Given a single SNPEffectMatrix
sem <- getSEMs(SEMC, "JUN")
plotSEM(sem)

# color by sequence
plotSEM(sem, motifSeq = "TGAGTCA", hindex = 2)
}
