% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSemMotifs.R
\name{plotSemMotifs}
\alias{plotSemMotifs}
\title{Plot non-alt versus alt binding propensity for a single variant}
\usage{
plotSemMotifs(
  semplObj,
  variant,
  label = "semId",
  changedCols = c("#F8766D", "dodgerblue2")
)
}
\arguments{
\item{semplObj}{a SemplScores object with scores populated}

\item{variant}{variant id to plot}

\item{label}{column in sem_metadata slot of semplObj to use for point labels}

\item{changedCols}{vector of length 2 with colors to use for plotting gained
and lost motifs respectively}
}
\value{
a \code{ggplot} scatter plot of non-alt binding propensity versus alt
binding propensity
}
\description{
Plot non-alt versus alt binding propensity for a single variant
}
