% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSemMotifs.R
\name{plotSemMotifs}
\alias{plotSemMotifs}
\title{Plot non-risk versus risk binding propensity for a single variant}
\usage{
plotSemMotifs(sempl_obj, variant, label = "semId")
}
\arguments{
\item{sempl_obj}{a SemplScores object with scores populated}

\item{variant}{variant id to plot}

\item{label}{column in sem_metadata slot of sempl_obj to use for point labels}
}
\value{
a \code{ggplot} scatter plot of non-risk binding propensity versus risk
binding propensity
}
\description{
Plot non-risk versus risk binding propensity for a single variant
}
