% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSemMotifs.R
\name{plotSemMotifs}
\alias{plotSemMotifs}
\title{Plot non-risk versus risk binding propensity}
\usage{
plotSemMotifs(sempl_obj, variant_i)
}
\arguments{
\item{sempl_obj}{a SemplR object with scores populated}

\item{variant_i}{numeric, index of the variant within the SEMplR object to plot}
}
\value{
a \code{ggplot} scatter plot of non-risk binding propensity versus risk
binding propensity
}
\description{
Plot non-risk versus risk binding propensity
}
