% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSemMotifs.R
\name{plotSemMotifs}
\alias{plotSemMotifs}
\title{Plot non-alt versus alt binding propensity for a single variant}
\usage{
plotSemMotifs(s, variant, label = "semId", cols = c("#F8766D", "dodgerblue2"))
}
\arguments{
\item{s}{a SemplScores object with scores populated}

\item{variant}{variant id to plot}

\item{label}{column in sem_metadata slot of semplObj to use for point labels}

\item{cols}{vector of length 2 with colors to use for plotting gained
and lost motifs respectively}
}
\value{
a \code{ggplot} scatter plot of non-alt binding propensity versus alt
binding propensity
}
\description{
Plot non-alt versus alt binding propensity for a single variant
}
\examples{
library(VariantAnnotation)
data(sc)

# create an SNP Effect Matrix (SEM)
sem <- matrix(rnorm(12), ncol = 4)
colnames(sem) <- c("A", "C", "G", "T")

# create a VRanges object
vr <- VRanges(seqnames = "chr12",
              ranges = 94136009, 
              ref = "G", alt = "C")

# calculate binding propensity
s <- scoreVariants(vr, sc, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)

plotSemMotifs(s, "chr12:94136009:G>C", label = "transcription_factor")

}
