% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSEMVariants.R
\name{plotSEMVariants}
\alias{plotSEMVariants}
\title{Plot non-alt versus alt binding propensity for a single motif}
\usage{
plotSEMVariants(
  s,
  sem,
  label = "varId",
  labsize = 4,
  cols = c("#F8766D", "dodgerblue2"),
  ptsize = 1
)
}
\arguments{
\item{s}{a SEMScores object with scores populated}

\item{sem}{a single character vector matching a SEM in the semplObj}

\item{label}{column in scores slot of semplObj to use for point labels}

\item{labsize}{numeric size of the point labels}

\item{cols}{vector of length 2 with colors to use for plotting gained
and lost motifs respectively}

\item{ptsize}{numeric size of the ggplot points}
}
\value{
a \code{ggplot} scatter plot of non-alt binding propensity versus alt
binding propensity
}
\description{
Plot non-alt versus alt binding propensity for a single motif
}
\examples{
library(VariantAnnotation)

# create a VRanges object
vr <- VRanges(
    seqnames = "chr12",
    ranges = 94136009,
    ref = "G", alt = "C"
)

# calculate binding propensity
s <- scoreVariants(vr, SEMC, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)

plotSEMVariants(s, "IKZF1_HUMAN.GM12878")

}
