% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSemVariants.R
\name{plotSemVariants}
\alias{plotSemVariants}
\title{Plot non-alt versus alt binding propensity for a single motif}
\usage{
plotSemVariants(
  semplObj,
  semId,
  label = "varId",
  changedCols = c("#F8766D", "dodgerblue2")
)
}
\arguments{
\item{semplObj}{a SemplScores object with scores populated}

\item{semId}{a single character vector matching a semId in the semplObj}

\item{label}{column in scores slot of semplObj to use for point labels}

\item{changedCols}{vector of length 2 with colors to use for plotting gained
and lost motifs respectively}
}
\value{
a \code{ggplot} scatter plot of non-alt binding propensity versus alt
binding propensity
}
\description{
Plot non-alt versus alt binding propensity for a single motif
}
