% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSemVariants.R
\name{plotSemVariants}
\alias{plotSemVariants}
\title{Plot non-alt versus alt binding propensity for a single motif}
\usage{
plotSemVariants(semplObj, semId, label = "varId")
}
\arguments{
\item{semplObj}{a SemplScores object with scores populated}

\item{semId}{numeric, index of the variant within the SEMplR object to plot}

\item{label}{column in scores slot of semplObj to use for point labels}
}
\value{
a \code{ggplot} scatter plot of non-alt binding propensity versus alt
binding propensity
}
\description{
Plot non-alt versus alt binding propensity for a single motif
}
