% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSemVariants.R
\name{plotSemVariants}
\alias{plotSemVariants}
\title{Plot non-risk versus risk binding propensity for a single motif}
\usage{
plotSemVariants(sempl_obj, sem_id, label = "varId")
}
\arguments{
\item{sempl_obj}{a SemplR object with scores populated}

\item{sem_id}{numeric, index of the variant within the SEMplR object to plot}

\item{label}{column in scores slot of sempl_obj to use for point labels}
}
\value{
a \code{ggplot} scatter plot of non-risk binding propensity versus risk
binding propensity
}
\description{
Plot non-risk versus risk binding propensity for a single motif
}
