% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSemVariants.R
\name{plotSemVariants}
\alias{plotSemVariants}
\title{Plot non-alt versus alt binding propensity for a single motif}
\usage{
plotSemVariants(
  semplObj,
  semId,
  label = "varId",
  changedCols = c("#F8766D", "dodgerblue2")
)
}
\arguments{
\item{semplObj}{a SemplScores object with scores populated}

\item{semId}{a single character vector matching a semId in the semplObj}

\item{label}{column in scores slot of semplObj to use for point labels}

\item{changedCols}{vector of length 2 with colors to use for plotting gained
and lost motifs respectively}
}
\value{
a \code{ggplot} scatter plot of non-alt binding propensity versus alt
binding propensity
}
\description{
Plot non-alt versus alt binding propensity for a single motif
}
\examples{
library(VariantAnnotation)
data(sc)

# create an SNP Effect Matrix (SEM)
sem <- matrix(rnorm(12), ncol = 4)
colnames(sem) <- c("A", "C", "G", "T")

# create a VRanges object
vr <- VRanges(seqnames = "chr12",
              ranges = 94136009, 
              ref = "G", alt = "C")

# calculate binding propensity
s <- scoreVariants(vr, sc, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)

plotSemVariants(s, "IKZF1_HUMAN.GM12878")

}
