% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreBinding.R
\name{scoreBinding}
\alias{scoreBinding}
\title{Calculate binding propensity for all SEM motifs and
genomic positions provided}
\usage{
scoreBinding(x, sem, genome, nFlank = NULL, seqId = NULL)
}
\arguments{
\item{x}{\code{GRanges} object or a vector of DNA sequences}

\item{sem}{A \code{SNPEffectMatrix} or \code{SNPEffectMatrixCollection} object}

\item{genome}{A \code{BSgenome} object for the genome build to use. ie.
\code{BSgenome.Hsapiens.UCSC.hg19::Hsapiens}. Required if providing a GRanges
object. Ignored if providing a vector of sequences.}

\item{nFlank}{Number of flanking nucleotides to add to provided range. By
default will add flank equal to the length of the longest motif. Ignored
if providing a vector of sequences.}

\item{seqId}{Column in \code{GRanges} object to use for unique id.
By default, ids will be generated from the \code{seqnames} and \code{ranges.}
Ignored if not providing a \code{GRanges} object.}
}
\value{
If a \code{GRanges} object is provided, return a \code{SEMplScores} object.
If a list of sequences is provided, just return the scoring table
}
\description{
Calculate binding propensity for all SEM motifs and
genomic positions provided
}
\examples{
library(GenomicRanges)

# load SEMs
data(sc)

# create a GRanges object
gr <- GenomicRanges::GRanges(seqnames = "chr12",
              ranges = 94136009)

# calculate binding propensity
scoreBinding(gr, sc, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)

}
