% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreBinding.R
\name{scoreBinding}
\alias{scoreBinding}
\title{Calculate binding propensity for all SEM motifs and
genomic positions provided}
\usage{
scoreBinding(x, semList, bs_genome_obj, allele = NULL)
}
\arguments{
\item{x}{\code{GRanges} object}

\item{semList}{A \code{SNPEffectMatrix} or \code{SNPEffectMatrixCollection} object}

\item{bs_genome_obj}{A \code{BSgenome} object for the genome build to use. ie.
\code{BSgenome.Hsapiens.UCSC.hg19::Hsapiens}}

\item{allele}{Column name in meta data storing allele}
}
\value{
a \code{data.table} object
}
\description{
Calculate binding propensity for all SEM motifs and
genomic positions provided
}
\examples{
library(GenomicRanges)

# create an SNP Effect Matrix (SEM)
s <- matrix(rnorm(12), ncol = 4)
colnames(s) <- c("A", "C", "G", "T")

# create a GRanges object
gr <- GenomicRanges::GRanges(seqnames = "chr12",
              ranges = 94136009, 
              ref = "G", alt = "C")

# create a list of SNPEffectMatrix objects
semList <- SNPEffectMatrix(s, baseline = -1, semId = "sem_id")

# calculate binding propensity
scoreBinding(gr, semList, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)

}
