% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SEMplScores.R
\name{scores}
\alias{scores}
\alias{scores,SEMScores-method}
\title{Accessor scores slot in a SEMScores object}
\usage{
scores(x)

\S4method{scores}{SEMScores}(x)
}
\arguments{
\item{x}{a SEMScores object}
}
\description{
Accessor scores slot in a SEMScores object
}
\examples{
library(VariantAnnotation)

# load default SEMs

# create a VRanges object
vr <- VRanges(
    seqnames = "chr12",
    ranges = 94136009,
    ref = "G", alt = "C"
)

# calculate binding propensity
s <- scoreVariants(vr, SEMC, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)

scores(s)

}
\keyword{internal}
