% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SemplScores.R
\name{scores}
\alias{scores}
\alias{scores,SemplScores-method}
\title{Accessor scores slot in a SemplScores object}
\usage{
scores(x)

\S4method{scores}{SemplScores}(x)
}
\arguments{
\item{x}{a SemplScores object}
}
\description{
Accessor scores slot in a SemplScores object
}
\examples{
library(VariantAnnotation)

# load default SEMs
data(sc)

# create a VRanges object
vr <- VRanges(seqnames = "chr12",
              ranges = 94136009, 
              ref = "G", alt = "C")

# calculate binding propensity
s <- scoreVariants(vr, sc, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)

scores(s)

}
\keyword{internal}
