% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SEMplScores.R
\name{scores}
\alias{scores}
\alias{scores,SEMplScores-method}
\title{Accessor scores slot in a SEMplScores object}
\usage{
scores(x)

\S4method{scores}{SEMplScores}(x)
}
\arguments{
\item{x}{a SEMplScores object}
}
\description{
Accessor scores slot in a SEMplScores object
}
\examples{
library(VariantAnnotation)

# load default SEMs
data(sc)

# create a VRanges object
vr <- VRanges(
    seqnames = "chr12",
    ranges = 94136009,
    ref = "G", alt = "C"
)

# calculate binding propensity
s <- scoreVariants(vr, sc, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)

scores(s)

}
\keyword{internal}
