% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SemplScores.R
\name{scores}
\alias{scores}
\alias{scores,SemplScores-method}
\title{Accessor scores slot in a SemplScores object}
\usage{
scores(x)

\S4method{scores}{SemplScores}(x)
}
\arguments{
\item{x}{a SemplScores object}
}
\description{
Accessor scores slot in a SemplScores object
}
\keyword{internal}
