% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreVariants.R
\name{scoreVariants}
\alias{scoreVariants}
\title{Calculate risk/non-risk binding propensity for all SEM motifs and
variants provided}
\usage{
scoreVariants(
  x,
  sem,
  genome,
  refCol = NULL,
  altCol = NULL,
  varId = NULL,
  rc = TRUE
)
}
\arguments{
\item{x}{\code{VRanges} object}

\item{sem}{a list of \code{SNPEffectMatrix} objects}

\item{genome}{A \code{BSgenome} object for the genome build to use. ie.
\code{BSgenome.Hsapiens.UCSC.hg19::Hsapiens}}

\item{refCol}{If providing a GRanges, the meta data column name with the
reference (ref) allele. Ignored if providing a VRanges object.}

\item{altCol}{If providing a GRanges, the meta data column name with the
alternative (alt) allele. Ignored if providing a VRanges object.}

\item{varId}{A column name in the meta data of x to use as a unique id.}

\item{rc}{plot the reverse complement SEMs}
}
\value{
a SEMScores object with slots for the ranges of the provided
variants with an added \code{sequence} column with the sequence scored,
the SEM metadata, and the resulting scoring table.
These slots are accessible with the \code{getRanges()}, \code{semData()}, and
\code{scores()} accessor functions respectively.

The scoring table will contain the following columns:
\itemize{
\item varId: a unique identifier for the sequence scored
\item SEM: the identifier for the SEM
\item rc: the orientation of the sequence (fwd or rev)
\item score: the unnormalized SEM score
\item scoreNorm: the SEM score normalized to it's corresponding baseline
\item index: the index of the motif with the highest SEM score within the
sequence scored
\item seq: the motif sequence with the highest SEM scores within the sequence
scored
}
}
\description{
Calculate risk/non-risk binding propensity for all SEM motifs and
variants provided
}
\examples{
library(VariantAnnotation)

# load default SEMs

# create a VRanges object
x <- VRanges(
    seqnames = "chr12",
    ranges = 94136009,
    ref = "G", alt = "C"
)

# calculate binding propensity
scoreVariants(x, SEMC, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)

}
