% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreVariants.R
\name{scoreVariants}
\alias{scoreVariants}
\title{Calculate risk/non-risk binding propensity for all SEM motifs and
variants provided}
\usage{
scoreVariants(x, sem, genome, refCol = NULL, altCol = NULL, varId = NULL)
}
\arguments{
\item{x}{\code{VRanges} object}

\item{sem}{a list of \code{SNPEffectMatrix} objects}

\item{genome}{A \code{BSgenome} object for the genome build to use. ie.
\code{BSgenome.Hsapiens.UCSC.hg19::Hsapiens}}

\item{refCol}{If providing a GRanges, the meta data column name with the
reference (ref) allele. Ignored if providing a VRanges object.}

\item{altCol}{If providing a GRanges, the meta data column name with the
alternative (alt) allele. Ignored if providing a VRanges object.}

\item{varId}{A column name in the meta data of x to use as a unique id.}
}
\value{
a SEMplScores object
}
\description{
Calculate risk/non-risk binding propensity for all SEM motifs and
variants provided
}
\examples{
library(VariantAnnotation)

# load default SEMs
data(SEMC)

# create a VRanges object
x <- VRanges(
    seqnames = "chr12",
    ranges = 94136009,
    ref = "G", alt = "C"
)

# calculate binding propensity
scoreVariants(x, SEMC, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)

}
