% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreVariants.R
\name{scoreVariants}
\alias{scoreVariants}
\title{Calculate risk/non-risk binding propensity for all SEM motifs and
variants provided}
\usage{
scoreVariants(vr, semList, bs_genome_obj)
}
\arguments{
\item{vr}{\code{VRanges} object}

\item{semList}{a list of \code{SNPEffectMatrix} objects}

\item{bs_genome_obj}{A \code{BSgenome} object for the genome build to use. ie.
\code{BSgenome.Hsapiens.UCSC.hg19::Hsapiens}}
}
\value{
a SemplScores object
}
\description{
Calculate risk/non-risk binding propensity for all SEM motifs and
variants provided
}
\examples{
library(VariantAnnotation)

# load default SEMs
data(sc)

# create a VRanges object
vr <- VRanges(seqnames = "chr12",
              ranges = 94136009, 
              ref = "G", alt = "C")

# calculate binding propensity
scoreVariants(vr, sc, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)

}
