% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semMotifBinding.R
\name{scoreVariants}
\alias{scoreVariants}
\title{Define a function to calculate risk/non-risk binding propensity
for all SEM motifs provided}
\usage{
scoreVariants(varId, varSeq, semId, semMtx, bl, offset)
}
\arguments{
\item{varId}{a list of unique character identifiers for each variant}

\item{varSeq}{a list of variant sequences}

\item{semId}{a unique character identifier for the sem}

\item{semMtx}{a numeric matrix of motif position by nucleic acid}

\item{bl}{a numeric value of the baseline to use for normalization}

\item{offset}{maximum number of offset basepairs offset from variant}
}
\description{
Define a function to calculate risk/non-risk binding propensity
for all SEM motifs provided
}
