% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semMotifBinding.R
\name{scoreVariants}
\alias{scoreVariants}
\title{Define a function to calculate risk/non-risk binding propensity
for all SEM motifs provided}
\usage{
scoreVariants(varId, varSeq, semObj, nflank)
}
\arguments{
\item{varId}{a list of unique character identifiers for each variant}

\item{varSeq}{a list of variant sequences}

\item{semObj}{SNPEffectMatrix object}

\item{nflank}{maximum number of offset basepairs offset from variant}
}
\description{
Define a function to calculate risk/non-risk binding propensity
for all SEM motifs provided
}
