% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{sem}
\alias{sem}
\alias{sem,SNPEffectMatrix-method}
\title{Accessor function to pull the matrix slot from a SNPEffectMatrix object}
\usage{
sem(x)

\S4method{sem}{SNPEffectMatrix}(x)
}
\arguments{
\item{x}{a SNPEffectMatrix object}
}
\description{
Accessor function to pull the matrix slot from a SNPEffectMatrix object
}
