% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-SNPEffectMatrixCollection.R, R/methods-SemplScores.R
\name{semData}
\alias{semData}
\alias{semData,SNPEffectMatrixCollection-method}
\alias{semData,SemplScores-method}
\title{Access semData from a SNPEffectMatrixCollection object}
\usage{
semData(x)

\S4method{semData}{SNPEffectMatrixCollection}(x)

\S4method{semData}{SemplScores}(x)
}
\arguments{
\item{x}{a SemplScores object}
}
\value{
A \code{data.table} is returned
}
\description{
Access semData from a SNPEffectMatrixCollection object

Accessor semData slot in a SemplScores object
}
\examples{
## Create example SEM
df <- data.frame(A=c(1, 2, 3),
                 C=c(1, 2, 3),
                 G=c(1, 2, 3),
                 T=c(1, 2, 3))
sem_data <- data.frame(SEM = "motif_id", TF = "tf_id")
s <- SNPEffectMatrix(df, 1.205, "motif_id")
sc <- SNPEffectMatrixCollection(s, semData = sem_data, semKey = "SEM")

## Access count matrix
semData(sc)

}
