% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{semData}
\alias{semData}
\alias{semData,SemplScores-method}
\title{Accessor semData slot in a SemplScores object}
\usage{
semData(x)

\S4method{semData}{SemplScores}(x)
}
\arguments{
\item{x}{a SemplScores object}
}
\description{
Accessor semData slot in a SemplScores object
}
