% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SEMplScores.R,
%   R/methods-SNPEffectMatrixCollection.R
\name{semData}
\alias{semData}
\alias{semData,SEMplScores-method}
\alias{semData,SNPEffectMatrixCollection-method}
\title{Accessor semData slot in a SEMplScores object}
\usage{
semData(x)

\S4method{semData}{SEMplScores}(x)

\S4method{semData}{SNPEffectMatrixCollection}(x)
}
\arguments{
\item{x}{SNPEffectMatrixCollection object}
}
\value{
A \code{data.table} is returned
}
\description{
Accessor semData slot in a SEMplScores object

Access semData from a SNPEffectMatrixCollection object
}
\examples{
library(VariantAnnotation)

# load default SEMs
data(sc)

# create a VRanges object
vr <- VRanges(
    seqnames = "chr12",
    ranges = 94136009,
    ref = "G", alt = "C"
)

# calculate binding propensity
s <- scoreVariants(vr, sc, BSgenome.Hsapiens.UCSC.hg19::Hsapiens)

semData(s)

## Create example SEM
df <- data.frame(
    A = c(1, 2, 3),
    C = c(1, 2, 3),
    G = c(1, 2, 3),
    T = c(1, 2, 3)
)
sem_data <- data.frame(SEM = "motif_id", TF = "tf_id")
s <- SNPEffectMatrix(df, 1.205, "motif_id")
sc <- SNPEffectMatrixCollection(s, semData = sem_data, semKey = "SEM")

## Access count matrix
semData(sc)

}
