% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{semId}
\alias{semId}
\alias{semId,SNPEffectMatrix-method}
\title{Accessor function to pull the semId slot from a SNPEffectMatrix object}
\usage{
semId(x)

\S4method{semId}{SNPEffectMatrix}(x)
}
\arguments{
\item{x}{a SNPEffectMatrix object}
}
\description{
Accessor function to pull the semId slot from a SNPEffectMatrix object
}
