% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semMotifBinding.R
\name{semMotifBinding}
\alias{semMotifBinding}
\title{Define a function to calculate risk/non-risk binding propensity
for all SEM motifs provided}
\usage{
semMotifBinding(vr, semList = NULL)
}
\arguments{
\item{vr}{\code{VRanges} object}

\item{semList}{a list of \code{SNPEffectMatrix} objects. Defaults to loading .sems
and baseline data from github.com/Boyle-Lab/SEMpl/raw/master/SEMs/}
}
\description{
Define a function to calculate risk/non-risk binding propensity
for all SEM motifs provided
}
