% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semMotifBinding.R
\name{semMotifBinding}
\alias{semMotifBinding}
\title{Define a function to calculate risk/non-risk binding propensity
for all SEM motifs provided}
\usage{
semMotifBinding(vr, semFiles, semBaselines)
}
\arguments{
\item{vr}{\code{VRanges} object}

\item{semFiles}{path to sem files}

\item{semBaselines}{path to sem baselines file}
}
\description{
Define a function to calculate risk/non-risk binding propensity
for all SEM motifs provided
}
