% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semMotifBinding.R
\name{semMotifBinding}
\alias{semMotifBinding}
\title{Define a function to calculate risk/non-risk binding propensity
for all SEM motifs provided}
\usage{
semMotifBinding(
  chr = "chr12",
  pos = 94136009,
  ref = "G",
  alt = "C",
  semFiles,
  semBaselines
)
}
\arguments{
\item{chr}{character of chromosome}

\item{pos}{integer of SNP position}

\item{ref}{character of reference nucleotide (can be upper or lower)}

\item{alt}{character of alternate nulceotide (can be upper or lower)}

\item{semFiles}{character vector of paths to each \code{.sem} file}

\item{semBaselines}{file path to sem baselines}
}
\description{
Define a function to calculate risk/non-risk binding propensity
for all SEM motifs provided
}
