% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SemplR}
\alias{SemplR}
\alias{SemplR-class}
\title{SemplR object and constructor}
\usage{
SemplR(variants, scores)
}
\arguments{
\item{variants}{A \code{VRanges} object to hold one or more variants}

\item{scores}{A \code{data.table} object for motif information and binding scores}
}
\value{
a SemplR object
}
\description{
Constructs a SemplR class object.
}
