% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SemplScores.R
\docType{class}
\name{SemplScores}
\alias{SemplScores}
\title{SemplScores object and constructor}
\usage{
SemplScores(variants = NULL, semData = NULL, scores = NULL)
}
\arguments{
\item{variants}{A \code{VRanges} object to hold one or more variants}

\item{semData}{A named list of SNPEffectMatrix objects}

\item{scores}{(optional) A \code{data.table} object for motif information and
binding scores}
}
\value{
a SemplScores object
}
\description{
Constructs a SemplScores class object.
}
\examples{
# load default SEMs
data(sc)

# create a VRanges object
vr <- VariantAnnotation::VRanges(seqnames = c("chr12", "chr19"),
                                 ranges = c(94136009, 10640062), 
                                 ref = c("G", "T"), alt = c("C", "A"))

SemplScores(variants = vr, semData = semData(sc))

}
