% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SemplScores}
\alias{SemplScores}
\alias{SemplScores-class}
\title{SemplScores object and constructor}
\usage{
SemplScores(variants = NULL, scores = NULL, sem_metadata = NULL)
}
\arguments{
\item{variants}{A \code{VRanges} object to hold one or more variants}

\item{scores}{A \code{data.table} object for motif information and binding scores}

\item{sem_metadata}{A \code{data.table} object of metadata for the SEMs with one
row for each SEM}
}
\value{
a SemplScores object
}
\description{
Constructs a SemplScores class object.
}
