% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SEMplScores.R
\docType{class}
\name{SEMplScores}
\alias{SEMplScores}
\title{SEMplScores object and constructor}
\usage{
SEMplScores(ranges = NULL, semData = NULL, scores = NULL)
}
\arguments{
\item{ranges}{A \code{GRanges} or \code{VRanges} object to hold one or more variants}

\item{semData}{A named list of SNPEffectMatrix objects}

\item{scores}{(optional) A \code{data.table} object for motif information and
binding scores}
}
\value{
a SEMplScores object
}
\description{
Constructs a SEMplScores class object.
}
\examples{
# load default SEMs
data(sc)

# create a VRanges object
vr <- VariantAnnotation::VRanges(seqnames = c("chr12", "chr19"),
                                 ranges = c(94136009, 10640062), 
                                 ref = c("G", "T"), alt = c("C", "A"))

SEMplScores(ranges = vr, semData = semData(sc))

}
