% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SEMplScores.R
\docType{class}
\name{SEMScores}
\alias{SEMScores}
\title{SEMScores object and constructor}
\usage{
SEMScores(ranges = NULL, semData = NULL, scores = NULL)
}
\arguments{
\item{ranges}{A \code{GRanges} or \code{VRanges} object to hold one or more variants}

\item{semData}{A named list of SNPEffectMatrix objects}

\item{scores}{(optional) A \code{data.table} object for motif information and
binding scores}
}
\value{
a SEMScores object
}
\description{
Constructs a SEMScores class object.
}
\examples{
# load default SEMs

# create a VRanges object
vr <- VariantAnnotation::VRanges(
    seqnames = c("chr12", "chr19"),
    ranges = c(94136009, 10640062),
    ref = c("G", "T"), alt = c("C", "A")
)

SEMScores(ranges = vr, semData = semData(SEMC))

}
