% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,SequenceFrame-method}
\alias{show,SequenceFrame-method}
\title{Show function for SequenceFrame}
\usage{
\S4method{show}{SequenceFrame}(object)
}
\arguments{
\item{object}{a SequenceFrame object}
}
\description{
Show function for SequenceFrame
}
