% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SEMplScores.R,
%   R/methods-SNPEffectMatrixCollection.R
\name{show,SEMplScores-method}
\alias{show,SEMplScores-method}
\alias{show,SNPEffectMatrixCollection-method}
\title{Show method for SEMplScores objects}
\usage{
\S4method{show}{SEMplScores}(object)

\S4method{show}{SNPEffectMatrixCollection}(object)
}
\arguments{
\item{object}{a SNPEffectMatrixCollection object}
}
\description{
Prints information about the number of variants, SEM meta data columns, and
the scoring table if scoreVariants has been run.

Prints information about the number of SEMs and meta data columns included
in the object.
}
