% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SNPEffectMatrix-class}
\alias{SNPEffectMatrix-class}
\title{https://doi.org/10.1093/bioinformatics/btz612
A SNP Effect Matrix (SEM) estimates the binding affinity of every possible
mutation in a particular transcription factor (TF) binding motif
This class contains three slots: the matrix, the baseline value, and
metadata, which contains the TF name and}
\description{
https://doi.org/10.1093/bioinformatics/btz612
A SNP Effect Matrix (SEM) estimates the binding affinity of every possible
mutation in a particular transcription factor (TF) binding motif
This class contains three slots: the matrix, the baseline value, and
metadata, which contains the TF name and
}
\section{Slots}{

\describe{
\item{\code{matrix}}{The SEM itself as a data table Rows represent sequence position
(variable length), columns represent effects due to each nucleotide base
A, C, G, T (fixed length: 4)}

\item{\code{baseline}}{A scrambled baseline, representing the binding score of
randomly scrambled kmers of the same length. This is the binding cutoff
for a TF}

\item{\code{tf_name}}{Name of the TF relevant to the SEM}

\item{\code{pwm_filename}}{Name of the original PWM file used to generate the SEM}
}}

