% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-SNPEffectMatrix.R
\docType{class}
\name{SNPEffectMatrix-class}
\alias{SNPEffectMatrix-class}
\alias{show,SNPEffectMatrix-method}
\title{SNP Effect Matrix (SEM)}
\usage{
\S4method{show}{SNPEffectMatrix}(object)
}
\arguments{
\item{object}{SNPEffectMatrix object.}
}
\description{
A SNP Effect Matrix (SEM) estimates the binding affinity of every possible
mutation in a particular transcription factor (TF) binding motif.
Read more about SEMs here: https://doi.org/10.1093/bioinformatics/btz612
This class contains three slots: the matrix, the baseline value, and
a unique id
}
\section{Slots}{

\describe{
\item{\code{sem}}{The SEM itself as a data table Rows represent sequence position
(variable length), columns represent effects due to each nucleotide base
A, C, G, T (fixed length: 4)}

\item{\code{baseline}}{A scrambled baseline, representing the binding score of
randomly scrambled kmers of the same length. This is the binding cutoff
for a TF}

\item{\code{semId}}{basename of the sem file}
}}

\examples{
# build a SEM example matrix
m <- matrix(rnorm(16), nrow = 4)
colnames(m) <- c("A", "C", "G", "T")

SNPEffectMatrix(sem = m, baseline = 1, semId = "sem_id")

}
