% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SNPEffectMatrix.R
\docType{class}
\name{SNPEffectMatrix}
\alias{SNPEffectMatrix}
\title{SNPEffectMatrix object and constructor}
\usage{
SNPEffectMatrix(sem, baseline, semId)
}
\arguments{
\item{sem}{A \code{data.table} object of format base position (rows) by nucleic
acid (columns). Column names must inclue A, C, T, and G; other columns
will be ignored.}

\item{baseline}{A \code{numeric} scrambled baseline, representing the binding
score of randomly scrambled kmers of the same length.}

\item{semId}{A \code{character} unique identifier for the SEM.}
}
\value{
a SNPEffectMatrix object
}
\description{
Constructs a SNPEffectMatrix class object.
}
