% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SNPEffectMatrix}
\alias{SNPEffectMatrix}
\alias{SNPEffectMatrix-class}
\title{SNPEffectMatrix object and constructor}
\usage{
SNPEffectMatrix(
  matrix,
  baseline,
  sem_id,
  tf_name = "",
  ensembl_id = "",
  uniprot_id = "",
  cell_type = ""
)
}
\arguments{
\item{matrix}{A \code{data.table} object to hold one or more variants}

\item{baseline}{A \code{numeric} scrambled baseline, representing the binding
score of randomly scrambled kmers of the same length.}

\item{sem_id}{basename of the sem file}

\item{tf_name}{(optional) Name of the TF relevant to the SEM}

\item{ensembl_id}{(optional) ENSEMBL gene id of the transcription factor}

\item{uniprot_id}{(optional) Uniprot protein id of the transcription factor}

\item{cell_type}{(optional) cell type/line used for ChipSeq experiment}
}
\value{
a SNPEffectMatrix object
}
\description{
Constructs a SNPEffectMatrix class object.
}
