% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SNPEffectMatrixCollection-class}
\alias{SNPEffectMatrixCollection-class}
\title{A collection of SNPEffectMatrix objects and their corresponding meta data}
\description{
A collection of SNPEffectMatrix objects and their corresponding meta data
}
\section{Slots}{

\describe{
\item{\code{sems}}{A list of SNPEffectMatrix objects}

\item{\code{semData}}{A \code{data.table} with meta data on each SEM}

\item{\code{semKey}}{The name of a column in semData that matches the semIds in the
sems list as a \code{character}. If column entries have a .sem suffix, a new
column named SEM_KEY will be created without the .sem suffixes.}
}}

\examples{
# build a SEM example matrix
m <- matrix(rnorm(16), nrow = 4)
colnames(m) <- c("A", "C", "G", "T")

# build a SNPEffectMatrix object
sm <- SNPEffectMatrix(sem = m, baseline = 1, semId = "sem_name")

# create a meta data table
md <- data.table::data.table(tf = "tf_name", sem = "sem_name")

# build a collection with 1 SEM
SNPEffectMatrixCollection(sems = list(sm), semData = md, semKey = "sem")

}
