% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SNPEffectMatrixCollection.R
\docType{class}
\name{SNPEffectMatrixCollection}
\alias{SNPEffectMatrixCollection}
\title{SNPEffectMatrixCollection object and constructor}
\usage{
SNPEffectMatrixCollection(sems, semData = NULL, semKey = "")
}
\arguments{
\item{sems}{A list of SNPEffectMatrix objects}

\item{semData}{A \code{data.table} containing meta data for each SEM}

\item{semKey}{A column name in semData corresponding to the semIds in each
SNPEffectMatrix object in the sems parameter}
}
\value{
a SNPEffectMatrixCollection object
}
\description{
Constructs a SNPEffectMatrixCollection class object.
}
