% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SemplScores.R
\name{variants}
\alias{variants}
\alias{variants,SemplScores-method}
\title{Accessor variants slot in a SemplScores object}
\usage{
variants(x)

\S4method{variants}{SemplScores}(x)
}
\arguments{
\item{x}{a SemplScores object}
}
\description{
Accessor variants slot in a SemplScores object
}
