% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewFrames.R
\name{viewFrames}
\alias{viewFrames}
\title{View the top scoring frames for a given variant and SEM}
\usage{
viewFrames(sempl_obj, score_index)
}
\arguments{
\item{sempl_obj}{SemplScores object}

\item{score_index}{index of score to view}
}
\value{
a SequenceFrame object
}
\description{
View the top scoring frames for a given variant and SEM
}
